refactor: consolidate theme handling logic and remove ThemeToggle component
